requirements.txt
aniso8601==8.1.1
aspy.refactor-imports==2.1.1
astroid==2.3.3
attrs==20.3.0
bleach==3.3.0
build==1.0.3
cached-property==1.5.2
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==8.1.7
cryptography==3.3.2
docutils==0.16
Flask==1.1.1
Flask-Cors==3.0.8
Flask-RESTful==0.3.8
hiredis==1.0.1
idna==2.10
importlib-metadata==6.8.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.3
keyring==22.0.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
marshmallow==3.10.0
marshmallow-dataclass==7.3.0
marshmallow-enum==1.5.1
marshmallow-union==0.1.15.post1
mccabe==0.6.1
more-itertools==8.7.0
mypy==0.761
mypy-extensions==0.4.3
packaging==20.9
pip-tools==7.3.0
pkginfo==1.7.0
pluggy==0.13.1
progress==1.5
py==1.10.0
pycparser==2.20
Pygments==2.7.4
pylint==2.4.4
pyparsing==2.4.7
pyproject_hooks==1.0.0
pytest==5.4.1
pytest-dotenv==0.5.2
python-decouple==3.8
python-dotenv==0.15.0
pytz==2021.1
readme-renderer==28.0
redis==3.5.0
redistimeseries==0.8.0
reorder-python-imports==1.9.0
requests==2.25.1
requests-toolbelt==0.9.1
rmtest==0.7.0
six==1.15.0
tomli==2.0.1
tqdm==4.56.0
twine==3.1.1
typed-ast==1.4.2
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.3
wcwidth==0.2.5
webargs==6.0.0
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.11.2
yapf==0.29.0
zipp==3.17.0
